fix: remove elon
