--- exercises/concept/elons-toy-car/.docs/introduction.md.tpl
@@ -1,3 +1,0 @@
-# Introduction
-
-%{concept:classes}
