--- exercises/concept/elons-toy-car/.meta/config.json
@@ -1,24 +1,0 @@
-{
-  "authors": [
-    "mikedamay"
-  ],
-  "contributors": [
-    "mirkoperillo"
-  ],
-  "files": {
-    "solution": [
-      "src/main/java/ElonsToyCar.java"
-    ],
-    "test": [
-      "src/test/java/ElonsToyCarTest.java"
-    ],
-    "exemplar": [
-      ".meta/src/reference/java/ElonsToyCar.java"
-    ],
-    "invalidator": [
-      "build.gradle"
-    ]
-  },
-  "icon": "elons-toys",
-  "blurb": "Learn about classes by working on a remote controlled car."
-}
