--- exercises/concept/elons-toy-car/src/main/java/ElonsToyCar.java
@@ -1,17 +1,0 @@
-public class ElonsToyCar {
-    public static ElonsToyCar buy() {
-        throw new UnsupportedOperationException("Please implement the (static) ElonsToyCar.buy()  method");
-    }
-
-    public String distanceDisplay() {
-        throw new UnsupportedOperationException("Please implement the ElonsToyCar.distanceDisplay()  method");
-    }
-
-    public String batteryDisplay() {
-        throw new UnsupportedOperationException("Please implement the ElonsToyCar.batteryDisplay()  method");
-    }
-
-    public void drive() {
-        throw new UnsupportedOperationException("Please implement the ElonsToyCar.drive()  method");
-    }
-}
