+++ exercises/concept/jedliks-toy-car/.docs/introduction.md.tpl
@@ -1,0 +1,3 @@
+# Introduction
+
+%{concept:classes}
