+++ exercises/concept/jedliks-toy-car/.meta/design.md
@@ -1,0 +1,30 @@
+# Design
+
+## Learning objectives
+
+- Know what classes are.
+- Know what encapsulation is.
+- Know what fields are.
+- Know how to create an object.
+- Know how to update state through methods.
+- Know about the `void` type.
+
+## Out of scope
+
+- Reference equality.
+- Constructors.
+- Interfaces.
+- Inheritance.
+- Finalizers.
+- Method overloading.
+
+## Concepts
+
+- `classes`: know what classes are; know what encapsulation is; know what fields are; know how to create an object; know how to update state through methods; know about the `void` type.
+
+## Prerequisites
+
+- `basics`: know how to define a basic class with basic methods.
+- `strings`: know how to do basic string interpolation.
+- `numbers`: know how to compare numbers.
+- `conditionals`: know how to do conditional logic.
