+++ exercises/concept/jedliks-toy-car/src/main/java/JedliksToyCar.java
@@ -1,0 +1,17 @@
+public class JedliksToyCar {
+    public static JedliksToyCar buy() {
+        throw new UnsupportedOperationException("Please implement the (static) JedliksToyCar.buy()  method");
+    }
+
+    public String distanceDisplay() {
+        throw new UnsupportedOperationException("Please implement the JedliksToyCar.distanceDisplay()  method");
+    }
+
+    public String batteryDisplay() {
+        throw new UnsupportedOperationException("Please implement the JedliksToyCar.batteryDisplay()  method");
+    }
+
+    public void drive() {
+        throw new UnsupportedOperationException("Please implement the JedliksToyCar.drive()  method");
+    }
+}
