+++ exercises/settings.gradle
@@ -8,7 +8,7 @@ include 'concept:booking-up-for-beauty'
 include 'concept:calculator-conundrum'
 include 'concept:captains-log'
 include 'concept:cars-assemble'
-include 'concept:elons-toy-car'
+include 'concept:jedliks-toy-car'
 include 'concept:football-match-reports'
 include 'concept:gotta-snatch-em-all'
 include 'concept:international-calling-connoisseur'
